generalize MultiIntervalFetcher and MultiIntervalExtractor
