+++ kipoiseq/extractors/__init__.py
@@ -1,4 +1,6 @@
 from .base import *
+from .fasta import *
+from .gtf import *
 from .vcf import *
 from .vcf_seq import *
 from .vcf_matching import *
