+++ kipoiseq/extractors/protein.py
@@ -2,12 +2,13 @@
 import abc
 
 import pandas as pd
-import numpy as np
+from kipoiseq.extractors.gtf import CDSFetcher, gtf_row2interval
 from kipoiseq.extractors.vcf_query import VariantIntervalQueryable
 
 from kipoiseq.dataclasses import Interval, Variant
 from kipoiseq.transforms.functional import rc_dna, translate
-from kipoiseq.extractors.base import FastaStringExtractor, BaseExtractor
+from kipoiseq.extractors.base import BaseExtractor, BaseMultiIntervalSeqExtractor, GenericMultiIntervalSeqExtractor
+from kipoiseq.extractors.fasta import FastaStringExtractor
 from kipoiseq.extractors.vcf import MultiSampleVCF
 from kipoiseq.extractors.vcf_seq import VariantSeqExtractor
 from kipoiseq.extractors.vcf_matching import SingleVariantMatcher, BaseVariantMatcher
@@ -31,6 +32,9 @@ def cut_transcript_seq(seq: str, tag: str):
     if ambiguous start and end or no tags provided, but the sequence has length % 3 != 0
     seq = 'NNN'
     """
+    # if not tag:
+    #     return seq
+
     if "cds_end_NF" in tag and "cds_start_NF" not in tag:
         # remove suffix
         seq_modulo = len(seq) % 3
@@ -56,365 +60,45 @@ def cut_transcript_seq(seq: str, tag: str):
     return seq
 
 
-def gtf_row2interval(row, interval_attrs: List[str] = None):
-    """
-    Convert gtf row object into interval class.
-    """
-    if interval_attrs:
-        interval_attrs = {i: row[i] for i in interval_attrs if i in row}
-    return Interval(str(row.Chromosome),
-                    int(row.Start),
-                    int(row.End),
-                    strand=str(row.Strand),
-                    attrs=interval_attrs)
-
-
-def _filter_valid_transcripts(gtf_df):
-    if 'transcript_support_level' in gtf_df:
-        gtf_df = gtf_df[~gtf_df.transcript_support_level.isnull()]
-        gtf_df = gtf_df[gtf_df.transcript_support_level != 'NA']
-        gtf_df.transcript_support_level = gtf_df.transcript_support_level.astype(
-            int)
-        gtf_df = gtf_df[~gtf_df.transcript_support_level.isna()]
-    else:
-        raise ValueError('Transcript support level not in gtf. '
-                         'Skipping the associated filters.')
-    return gtf_df
-
-
-def _get_biotype_str(df):
-    if 'transcript_biotype' in df:
-        return 'transcript_biotype'
-    elif 'gene_biotype' in df:
-        return 'gene_biotype'
-    else:
-        raise ValueError('Cannot obtain `biotype_str` from gtf file')
-
-
-def _filter_biotype(gtf_df, value):
-    """
-    Gets the biotype column and checks whether it equals `value`.
-    :param gtf_df: genome annotation GTF dataframe
-    :param value: The value to check for
-    :return: filtered dataframe
-    """
-    biotype_str = _get_biotype_str(gtf_df)
-    return gtf_df.query(
-        "{key} == '{value}'".format(key=biotype_str, value=value)
-    )
-
-
-def _filter_biotype_proteincoding(gtf_df):
-    return _filter_biotype(gtf_df, value="protein_coding")
-
-
-def _filter_tag(gtf_df, regex_contains):
-    return gtf_df.query(
-        "tag.notnull() & tag.str.contains('{}')".format(regex_contains)
-    )
-
-
-class SpliceIntervalFetcher(metaclass=abc.ABCMeta):
-
-    @abc.abstractmethod
-    def keys(self) -> List[object]:
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def get_intervals(self, key) -> List[Interval]:
-        """
-        Returns a list of Intervals on request of some key
-        :param key: The identifier for the requested intervals
-        :return: list of intervals
-        """
-        raise NotImplementedError()
-
-    def sel(self, *args, **kwargs):
-        """
-        alias for get_intervals
-        """
-        return self.get_intervals(*args, **kwargs)
-
-    def isel(self, idx):
-        """
-        Get interval by index
-        :param idx: index in the range (0, len(self))
-        """
-        return self.get_intervals(self.keys()[idx])
-
-    def __len__(self):
-        return len(self.keys())
-
-    def __getitem__(self, idx):
-        return self.isel(idx)
-
-    def __iter__(self):
-        for i in self.keys():
-            yield self.get_intervals(i)
-
-
-class CDSFetcher(SpliceIntervalFetcher):
-
-    def __init__(
-            self,
-            gtf_file,
-            filter_valid_transcripts=True,
-            filter_biotype=True,
-            filter_tag=True,
-            on_error_warn=True,
-    ):
-        """
-        Protein sequences in the genome
-        :param gtf_file:
-        """
-        self.gtf_file = str(gtf_file)
-        self.cds = self._read_cds(
-            self.gtf_file,
-            filter_valid_transcripts=filter_valid_transcripts,
-            filter_biotype=filter_biotype,
-            filter_tag=filter_tag,
-            on_error_warn=on_error_warn
-        )
-        self.transcripts = np.asarray(self.cds.index.unique())
-        self.on_error_warn = on_error_warn
-
-    @staticmethod
-    def _read_cds(
-            gtf_file,
-            filter_valid_transcripts=False,
-            filter_biotype=False,
-            filter_tag=False,
-            duplicate_attr=None,
-            on_error_warn=True,
-    ):
-        """
-        Read, extract and filter valid cds from the given gtf_file
-        :param gtf_file: path to the GTF file
-        """
-        import pyranges
-
-        if duplicate_attr == None:
-            # One row in the GTF file can have multiple tags;
-            # therefore, to filter them we have to allow duplicate attrs.
-            duplicate_attr = filter_tag
-
-        df = pyranges.read_gtf(gtf_file, as_df=True, duplicate_attr=duplicate_attr)
-
-        cds = CDSFetcher.get_cds_from_gtf(
-            df,
-            filter_valid_transcripts=filter_valid_transcripts,
-            filter_biotype=filter_biotype,
-            filter_tag=filter_tag,
-            on_error_warn=on_error_warn
-        )
-
-        cds = cds.set_index("transcript_id")
-        return cds
-
-    @staticmethod
-    def get_cds_from_gtf(
-            df,
-            filter_valid_transcripts=False,
-            filter_biotype=False,
-            filter_tag=False,
-            on_error_warn=True,
-    ):
-        """
-        Create DataFrame with valid cds
-
-        :param df: the GTF dataframe
-        :param filter_valid_transcripts: Filter for "transcript_support_level" column in GTF
-        :param filter_biotype: Filter for "[gene|transcript]_biotype == 'protein_coding'" in GTF
-        :param filter_tag: Filter for 'basic' or 'CCDS' tag
-        """
-        cds = df.query("(Feature == 'CDS') | (Feature == 'CCDS')")
-
-        if filter_biotype:
-            try:
-                cds = _filter_biotype_proteincoding(cds)
-            except ValueError as e:
-                if on_error_warn:
-                    log.warning("Error during filtering biotype: %s", e)
-                else:
-                    raise e
-        if filter_valid_transcripts:
-            try:
-                cds = _filter_valid_transcripts(cds)
-            except ValueError as e:
-                if on_error_warn:
-                    log.warning("Error during filtering for valid transcripts: %s", e)
-                else:
-                    raise e
-        if filter_tag:
-            try:
-                cds = _filter_tag(cds, regex_contains='basic|CCDS')
-            except ValueError as e:
-                if on_error_warn:
-                    log.warning("Error during filtering for tag: %s", e)
-                else:
-                    raise e
-
-        return cds
-
-    def get_intervals(self, transcript_id: str):
-        """
-        Create Interval objects of the cds for given transcript_id
-        """
-        cds = self.cds.loc[[transcript_id]]
-
-        check_strand = np.all(cds.iloc[0].Strand == cds.Strand)
-        if not check_strand:
-            if self.on_error_warn:
-                log.warning("Error while processing transcript '%s': strands of CDS parts differ", transcript_id)
-            else:
-                raise ValueError("Error while processing transcript '%s': strands of CDS parts differ" % transcript_id)
-
-        return [
-            gtf_row2interval(row, ['tag'])
-            for i, row in cds.sort_values("Start").iterrows()
-        ]
-
-    def keys(self):
-        return self.transcripts
-
-
-class UTRFetcher(SpliceIntervalFetcher):
-
-    def __init__(
-            self,
-            gtf_file,
-            feature_type="5UTR",
-            infer_from_cds=False,
-            on_error_warn=True,
-    ):
-        """
-        Protein sequences in the genome
-        :param gtf_file: path to the GTF file
-        :param feature_type: type of the feature that will be filtered for. In general '5UTR' or '3UTR'.
-        :param infer_from_cds: Substract the CDS from the exon regions to infer the UTR regions.
-            Will use 'feature_type' to decide whether '5UTR' or '3UTR' should be returned.
-        :param on_error_warn: Do not break on error; instead throw warning.
-        """
-        self.gtf_file = str(gtf_file)
-        self.utr_df = self._read_utr(
-            self.gtf_file,
-            feature_type=feature_type,
-            infer_from_cds=infer_from_cds,
-            on_error_warn=on_error_warn
-        )
-        self.transcripts = np.asarray(self.utr_df.index.unique())
-
-        # self.feature_type = feature_type
-        # self.infer_from_cds = infer_from_cds
-        self.on_error_warn = on_error_warn
-
-    @staticmethod
-    def _read_utr(
-            gtf_file,
-            feature_type="5UTR",
-            infer_from_cds=False,
-            on_error_warn=True,
-    ):
-        """
-        Read, extract and filter valid UTRs from the given gtf_file
-        :param gtf_file: path to the GTF file
-        :param feature_type: type of the feature that will be filtered for. In general '5UTR' or '3UTR'.
-        :param infer_from_cds: Substract the CDS from the exon regions to infer the UTR regions.
-            Will use 'feature_type' to decide whether '5UTR' or '3UTR' should be returned.
-        :param on_error_warn: Do not break on error; instead throw warning.
-        """
-        import pyranges
-
-        df = pyranges.read_gtf(gtf_file, as_df=True)
-
-        utr_df = UTRFetcher.get_utr_from_gtf(
-            df,
-            feature_type=feature_type,
-            infer_from_cds=infer_from_cds,
-            on_error_warn=on_error_warn
-        )
-
-        utr_df = utr_df.set_index("transcript_id")
-        return utr_df
-
-    @staticmethod
-    def get_utr_from_gtf(
-            df,
-            feature_type="5UTR",
-            infer_from_cds=False,
-            on_error_warn=True,
-    ):
-        """
-        Create DataFrame with valid UTRs
-        :param df: the GTF dataframe
-        :param feature_type: type of the feature that will be filtered for. In general '5UTR' or '3UTR'.
-        :param infer_from_cds: Substract the CDS from the exon regions to infer the UTR regions.
-            Will use 'feature_type' to decide whether '5UTR' or '3UTR' should be returned.
-        :param on_error_warn: Do not break on error; instead throw warning.
-        """
-        if infer_from_cds:
-            cds = CDSFetcher.get_cds_from_gtf(df=df, on_error_warn=on_error_warn)
-
-            # TODO: implement
-            if feature_type.upper() == "5UTR":
-                pass
-            if feature_type.upper() == "3UTR":
-                pass
-
-            raise NotImplementedError()
-        else:
-            utr_df = df.query("Feature == '{feature_type}')".format(feature_type=feature_type))
-
-        return utr_df
-
-    def get_intervals(self, transcript_id: str):
-        """
-        Create Interval objects of the cds for given transcript_id
-        """
-        utr_df = self.utr_df.loc[[transcript_id]]
-
-        check_strand = np.size(np.unique(utr_df.Strand)) == 1
-        if not check_strand:
-            if self.on_error_warn:
-                log.warning("Error while processing transcript '%s': strands of CDS parts differ", transcript_id)
-            else:
-                raise ValueError("Error while processing transcript '%s': strands of CDS parts differ" % transcript_id)
-
-        return [
-            gtf_row2interval(row, ['tag']) for i, row in utr_df.sort_values("Start").iterrows()
-        ]
-
-    def keys(self):
-        return self.transcripts
-
-
-class TranscriptSeqExtractor:
+class TranscriptSeqExtractor(GenericMultiIntervalSeqExtractor):
 
     def __init__(self, gtf_file, fasta_file):
         self.fasta_file = str(fasta_file)
         self.gtf_file = str(gtf_file)
-        self.fasta = FastaStringExtractor(self.fasta_file, use_strand=False)
-        self.cds_fetcher = CDSFetcher(self.gtf_file)
-        self.transcripts = self.cds_fetcher.transcripts
 
-    def __len__(self):
-        return len(self.cds_fetcher)
+        cds_fetcher = CDSFetcher(self.gtf_file)
+        self.cds = cds_fetcher.df
 
-    @staticmethod
-    def _prepare_seq(seqs: List[str], reverse_complement: Union[str, bool], tag: str):
+        # self.transcripts = self.cds_fetcher.keys()
+        extractor = FastaStringExtractor(self.fasta_file, use_strand=False)
+
+        super().__init__(
+            extractor=extractor,
+            interval_fetcher=cds_fetcher
+        )
+
+    @classmethod
+    def _prepare_seq(
+            cls,
+            seqs: List[str],
+            reverse_complement: Union[str, bool],
+            tag=None,
+            **kwargs
+    ) -> str:
         """
         Prepare the dna sequence in the final variant, which should be
         translated in amino acid sequence
         :param seqs: current dna sequence
         :param reverse_complement: should the dna be reverse-complemented?
-        :param tag: tags, which contain information about ambiguous start/end
-        :return: prepared dna sequence ready for translation into amino acid
-        sequence
+        :param tag: string containing tag list
+        :return: prepared dna sequence ready for translation into amino acid sequence
         """
-        seq = "".join(seqs)
-        if reverse_complement is True or reverse_complement == "-":
-            # optionally reverse complement
-            seq = rc_dna(seq)
+        seq = super()._prepare_seq(
+            seqs=seqs,
+            # intervals=intervals,
+            reverse_complement=reverse_complement,
+        )
+        # tag = intervals[0].attrs["tag"]
         seq = cut_transcript_seq(seq, tag)
         return seq
 
@@ -425,13 +109,36 @@ class TranscriptSeqExtractor:
         :param transcript_id:
         :return: dna sequence for the given transcript_id
         """
-        cds = self.cds_fetcher.get_intervals(transcript_id)
-        seqs = [self.fasta.extract(i) for i in cds]
+        return self.sel(transcript_id)
+        # cds_intervals = self.cds_fetcher.get_intervals(transcript_id)
+        # return self.extract(cds_intervals, tag=cds_intervals[0].attrs["tag"])
 
-        return self._prepare_seq(seqs, cds[0].strand, cds[0].attrs["tag"])
+        # return self._prepare_seq(seqs, cds_intervals[0].strand, cds_intervals[0].attrs["tag"])
 
-    def __getitem__(self, idx):
-        return self.get_seq(self.transcripts[idx])
+    def extract(self, intervals: List[Interval], **kwargs) -> str:
+        """
+        Extract and concatenate the sequence for a list of intervals
+        :param intervals: list of intervals
+        :param kwargs: will be passed on to `_prepare_seq()`
+        :return: concatenated dna sequence of the intervals
+        """
+        seqs = [self.extractor.extract(i) for i in intervals]
+
+        reverse_strand = False
+        if self.use_strand:
+            reverse_strand = intervals[0].neg_strand
+            if self.extractor.use_strand:
+                # If the fasta extractor already does the reversion, we do not have to do it manually.
+                # Instead, we need to reverse the list of sequences.
+                seqs.reverse()
+                reverse_strand = False
+
+        tag = intervals[0].attrs["tag"]
+
+        return self._prepare_seq(seqs, reverse_strand, tag=tag, **kwargs)
+
+    # def __getitem__(self, idx):
+    #     return self.get_seq(self.transcripts[idx])
 
     def overlaped_cds(self, variants):
         """
@@ -465,16 +172,16 @@ class TranscriptSeqExtractor:
 
 class ProteinSeqExtractor(TranscriptSeqExtractor):
 
-    @staticmethod
-    def _prepare_seq(seqs: List[str], reverse_complement: bool, tag: str):
+    @classmethod
+    def _prepare_seq(cls, *args, **kwargs):
         """
         Prepare the dna sequence and translate it into amino acid sequence
         :param seqs: current dna sequence
-        :param reverse_complement: sould the DNA be reverse-complemented?
-        :param tag: tags, which contain information about ambiguous start/end
+        :param reverse_complement: should the dna be reverse-complemented?
+        :param tag: string containing tag list
         :return: amino acid sequence
         """
-        return translate(TranscriptSeqExtractor._prepare_seq(seqs, reverse_complement, tag), True)
+        return translate(super()._prepare_seq(*args, **kwargs), True)
 
 
 class VCFSeqExtractor(metaclass=abc.ABCMeta):
@@ -501,7 +208,7 @@ class VCFSeqExtractor(metaclass=abc.ABCMeta):
             concat_by: Union[str, List[str]] = "transcript_id",
             **kwargs
     ):
-        return cls(ranges_df=pyranges.df.set_index(concat_by), **kwargs)
+        return cls(ranges_df=pyranges.cds.set_index(concat_by), **kwargs)
 
     def keys(self):
         return self.ranges_df.index.drop_duplicates()
@@ -517,7 +224,12 @@ class VCFSeqExtractor(metaclass=abc.ABCMeta):
         return [i.unstrand() for i in intervals]
 
     @classmethod
-    def _prepare_seq(cls, seqs: List[str], reverse_complement: bool) -> str:
+    def _prepare_seq(
+            cls,
+            seqs: List[str],
+            reverse_complement: Union[str, bool],
+            **kwargs
+    ) -> str:
         """
         Prepare the dna sequence in the final variant by concatenating and eventually reverse-complementing it
 
@@ -555,7 +267,7 @@ class VCFSeqExtractor(metaclass=abc.ABCMeta):
         iter_seqs = self.extract_query(variant_interval_queryable, sample_id=sample_id)
         for seqs, variant_info in iter_seqs:
             # 1st seq, 2nd variant info
-            yield self._prepare_seq(seqs, reverse_complement), variant_info
+            yield self._prepare_seq(seqs, intervals, reverse_complement), variant_info
 
         return None
 
@@ -612,7 +324,7 @@ class ProteinVCFSeqExtractor(VCFSeqExtractor, metaclass=abc.ABCMeta):
         reference_seq = FastaStringExtractor(self.fasta_file)
         multi_sample_VCF = MultiSampleVCF(self.vcf_file)
 
-        ranges_df = CDSFetcher(self.gtf_file).cds
+        ranges_df = CDSFetcher(self.gtf_file).df
 
         # dataframe to pyranges
         import pyranges
@@ -669,7 +381,7 @@ class ProteinVCFSeqExtractor(VCFSeqExtractor, metaclass=abc.ABCMeta):
             **kwargs
     ):
         reverse_complement = intervals[0].neg_strand
-        tags = intervals[0].attrs["tag"]
+        tag = intervals[0].attrs["tag"]
         # remove strand information
         intervals = [i.unstrand() for i in intervals]
 
@@ -678,7 +390,7 @@ class ProteinVCFSeqExtractor(VCFSeqExtractor, metaclass=abc.ABCMeta):
         iter_seqs = self.extract_query(variant_interval_queryable, sample_id=sample_id)
         for seqs, variant_info in iter_seqs:
             # 1st seq, 2nd variant info
-            yield ProteinSeqExtractor._prepare_seq(seqs, reverse_complement, tags), variant_info
+            yield ProteinSeqExtractor._prepare_seq(seqs, reverse_complement, tag=tag), variant_info
 
     def _filter_snv(self, variants):
         for variant in variants:
